fix(visualization): modify the image font loading logic (#2783)
fix(visualization): modify the image font loading logic to ensure the right font is installed

This commit updates the `pyproject.toml` to include Pillow version 10.1.0 as a dependency, which is necessary for supporting the size parameter in font loading. Additionally, the image font loading logic in `functional.py` has been modified to utilize the size parameter when loading the default font, enhancing text rendering capabilities in images.



Closes #2612

Signed-off-by: samet-akcay <samet.akcay@intel.com>